# frontend/openapi/events.json
{"openapi":"3.1.0","info":{"title":"Events Service","description":"Microservice de gestion des événements du DIT","version":"1.0.0"},"paths":{"/api/events":{"get":{"tags":["events"],"summary":"Lister les événements","description":"Liste les événements, avec filtres optionnels par date et par lieu.","operationId":"list_events","parameters":[{"name":"date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date"}},{"name":"location","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EventOut"},"title":"Response List Events"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["events"],"summary":"Créer un événement","operationId":"create_event","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/events/{event_id}":{"get":{"tags":["events"],"summary":"Détail d'un événement","operationId":"get_event","parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"integer","title":"Event Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["events"],"summary":"Modifier un événement","operationId":"update_event","parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"integer","title":"Event Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["events"],"summary":"Supprimer un événement","operationId":"delete_event","parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"integer","title":"Event Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/events/{event_id}/availability":{"get":{"tags":["events"],"summary":"Places restantes","operationId":"availability","parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"integer","title":"Event Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AvailabilityOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/events/{event_id}/seats/reserve":{"post":{"tags":["events"],"summary":"[interne] Réserver une place","operationId":"reserve_seat","parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"integer","title":"Event Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AvailabilityOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/events/{event_id}/seats/release":{"post":{"tags":["events"],"summary":"[interne] Libérer une place","operationId":"release_seat","parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"integer","title":"Event Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AvailabilityOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["health"],"summary":"Health","description":"Liveness : le processus répond. Utilisé par le HEALTHCHECK Docker.","operationId":"health","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health/ready":{"get":{"tags":["health"],"summary":"Readiness","description":"Readiness : le service ET sa base de données répondent.","operationId":"readiness","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AvailabilityOut":{"properties":{"event_id":{"type":"integer","title":"Event Id"},"capacity":{"type":"integer","title":"Capacity"},"seats_taken":{"type":"integer","title":"Seats Taken"},"seats_available":{"type":"integer","title":"Seats Available"},"is_full":{"type":"boolean","title":"Is Full"}},"type":"object","required":["event_id","capacity","seats_taken","seats_available","is_full"],"title":"AvailabilityOut"},"EventCreate":{"properties":{"title":{"type":"string","maxLength":200,"minLength":3,"title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"starts_at":{"type":"string","format":"date-time","title":"Starts At"},"location":{"type":"string","maxLength":200,"minLength":2,"title":"Location"},"capacity":{"type":"integer","maximum":10000.0,"exclusiveMinimum":0.0,"title":"Capacity"}},"type":"object","required":["title","starts_at","location","capacity"],"title":"EventCreate"},"EventOut":{"properties":{"title":{"type":"string","maxLength":200,"minLength":3,"title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"starts_at":{"type":"string","format":"date-time","title":"Starts At"},"location":{"type":"string","maxLength":200,"minLength":2,"title":"Location"},"capacity":{"type":"integer","maximum":10000.0,"exclusiveMinimum":0.0,"title":"Capacity"},"id":{"type":"integer","title":"Id"},"seats_taken":{"type":"integer","title":"Seats Taken"},"created_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["title","starts_at","location","capacity","id","seats_taken"],"title":"EventOut"},"EventUpdate":{"properties":{"title":{"anyOf":[{"type":"string","maxLength":200,"minLength":3},{"type":"null"}],"title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"starts_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Starts At"},"location":{"anyOf":[{"type":"string","maxLength":200,"minLength":2},{"type":"null"}],"title":"Location"},"capacity":{"anyOf":[{"type":"integer","maximum":10000.0,"exclusiveMinimum":0.0},{"type":"null"}],"title":"Capacity"}},"type":"object","title":"EventUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}